+++ packages/jest-config/src/Defaults.ts
@@ -6,6 +6,7 @@
  */
 
 import {sep} from 'path';
+import isCI = require('is-ci');
 import type {Config} from '@jest/types';
 import {replacePathSepForRegex} from 'jest-regex-util';
 import {NODE_MODULES} from './constants';
@@ -19,11 +20,14 @@ const defaultOptions: Config.DefaultOptions = {
   cache: true,
   cacheDirectory: getCacheDirectory(),
   changedFilesWithAncestor: false,
+  ci: isCI,
   clearMocks: false,
   collectCoverage: false,
   coveragePathIgnorePatterns: [NODE_MODULES_REGEXP],
   coverageProvider: 'babel',
   coverageReporters: ['json', 'text', 'lcov', 'clover'],
+  detectLeaks: false,
+  detectOpenHandles: false,
   errorOnDeprecated: false,
   expand: false,
   extensionsToTreatAsEsm: [],
@@ -34,6 +38,7 @@ const defaultOptions: Config.DefaultOptions = {
     throwOnModuleCollision: false,
   },
   injectGlobals: true,
+  listTests: false,
   maxConcurrency: 5,
   maxWorkers: '50%',
   moduleDirectories: ['node_modules'],
@@ -43,6 +48,7 @@ const defaultOptions: Config.DefaultOptions = {
   noStackTrace: false,
   notify: false,
   notifyMode: 'failure-change',
+  passWithNoTests: false,
   prettierPath: 'prettier',
   resetMocks: false,
   resetModules: false,