+++ packages/jest-types/src/Config.ts
@@ -57,11 +57,14 @@ export type DefaultOptions = {
   cache: boolean;
   cacheDirectory: Path;
   changedFilesWithAncestor: boolean;
+  ci: boolean;
   clearMocks: boolean;
   collectCoverage: boolean;
   coveragePathIgnorePatterns: Array<string>;
   coverageReporters: Array<CoverageReporterName>;
   coverageProvider: CoverageProvider;
+  detectLeaks: boolean;
+  detectOpenHandles: boolean;
   errorOnDeprecated: boolean;
   expand: boolean;
   extensionsToTreatAsEsm: Array<Path>;
@@ -69,6 +72,7 @@ export type DefaultOptions = {
   globals: ConfigGlobals;
   haste: HasteConfig;
   injectGlobals: boolean;
+  listTests: boolean;
   maxConcurrency: number;
   maxWorkers: number | string;
   moduleDirectories: Array<string>;
@@ -78,6 +82,7 @@ export type DefaultOptions = {
   noStackTrace: boolean;
   notify: boolean;
   notifyMode: NotifyMode;
+  passWithNoTests: boolean;
   prettierPath: string;
   resetMocks: boolean;
   resetModules: boolean;
@@ -121,6 +126,7 @@ export type InitialOptions = Partial<{
   bail: boolean | number;
   cache: boolean;
   cacheDirectory: Path;
+  ci: boolean;
   clearMocks: boolean;
   changedFilesWithAncestor: boolean;
   changedSince: string;